+++ spinetoolbox/project_items/exporter/ui/parameter_index_settings_window.py
@@ -13,7 +13,7 @@
 # Form implementation generated from reading ui file 'C:\data\GIT\SPINETOOLBOX\bin\..\spinetoolbox\project_items\exporter\ui\parameter_index_settings_window.ui',
 # licensing of 'C:\data\GIT\SPINETOOLBOX\bin\..\spinetoolbox\project_items\exporter\ui\parameter_index_settings_window.ui' applies.
 #
-# Created: Tue Feb 11 11:01:13 2020
+# Created: Thu Feb 13 11:53:50 2020
 #      by: pyside2-uic  running on PySide2 5.11.2
 #
 # WARNING! All changes made in this file will be lost!