+++ spinetoolbox/ui/duration_editor.py
@@ -13,7 +13,7 @@
 # Form implementation generated from reading ui file 'C:\data\GIT\SPINETOOLBOX\bin\..\spinetoolbox\ui\duration_editor.ui',
 # licensing of 'C:\data\GIT\SPINETOOLBOX\bin\..\spinetoolbox\ui\duration_editor.ui' applies.
 #
-# Created: Tue Feb 11 11:01:22 2020
+# Created: Thu Feb 13 11:53:58 2020
 #      by: pyside2-uic  running on PySide2 5.11.2
 #
 # WARNING! All changes made in this file will be lost!