+++ spinetoolbox/ui/plain_parameter_value_editor.py
@@ -13,7 +13,7 @@
 # Form implementation generated from reading ui file 'C:\data\GIT\SPINETOOLBOX\bin\..\spinetoolbox\ui\plain_parameter_value_editor.ui',
 # licensing of 'C:\data\GIT\SPINETOOLBOX\bin\..\spinetoolbox\ui\plain_parameter_value_editor.ui' applies.
 #
-# Created: Tue Feb 11 11:01:32 2020
+# Created: Thu Feb 13 11:54:09 2020
 #      by: pyside2-uic  running on PySide2 5.11.2
 #
 # WARNING! All changes made in this file will be lost!